CAP serializing 29012013

+++ map/map-api/src/main/java/org/mobicents/protocols/ss7/map/api/primitives/AlertingLevel.java
@@ -1,24 +1,25 @@
 /*
- * JBoss, Home of Professional Open Source
- * Copyright 2011, Red Hat, Inc. and/or its affiliates, and individual
- * contributors as indicated by the @authors tag. All rights reserved.
- * See the copyright.txt in the distribution for a full listing
- * of individual contributors.
- * 
- * This copyrighted material is made available to anyone wishing to use,
- * modify, copy, or redistribute it subject to the terms and conditions
- * of the GNU General Public License, v. 2.0.
- * 
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of 
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU 
- * General Public License for more details.
- * 
- * You should have received a copy of the GNU General Public License,
- * v. 2.0 along with this distribution; if not, write to the Free 
- * Software Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston,
- * MA 02110-1301, USA.
+ * TeleStax, Open Source Cloud Communications  Copyright 2012.
+ * and individual contributors
+ * by the @authors tag. See the copyright.txt in the distribution for a
+ * full listing of individual contributors.
+ *
+ * This is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU Lesser General Public License as
+ * published by the Free Software Foundation; either version 2.1 of
+ * the License, or (at your option) any later version.
+ *
+ * This software is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public
+ * License along with this software; if not, write to the Free
+ * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
+ * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
+
 package org.mobicents.protocols.ss7.map.api.primitives;
 
 /**
@@ -26,19 +27,19 @@ package org.mobicents.protocols.ss7.map.api.primitives;
  * 
  */
 public enum AlertingLevel {
-	Level0((byte)0), Level1((byte)1), Level2((byte)2);
+	Level0(0), Level1(1), Level2(2);
 
-	private final byte level;
+	private final int level;
 
-	private AlertingLevel(byte level) {
+	private AlertingLevel(int level) {
 		this.level = level;
 	}
 
-	public byte getLevel() {
+	public int getLevel() {
 		return this.level;
 	}
 
-	public static AlertingLevel getInstance(byte data) {
+	public static AlertingLevel getInstance(int data) {
 		switch (data) {
 		case 0:
 			return Level0;
